feat: add definition navigation via arrow keys
this commit adds navigation through a word's various definitions
with the left and right or up and down arrow keys.  It displays
the word, as well as the usage of the word's current definition (noun,
verb, adverb, etc.)

Allows user to 'keep' a definition or pick it by hitting the 'K' key.

Currently does not create image for upload.